Several pedantic warnings squished
git-svn-id: svn://svn.gna.org/svn/mysqlpp/trunk@2635 5946f024-35f4-0310-b1b3-9a3be3380cfb

+++ examples/multiquery.cpp
@@ -95,7 +95,7 @@ print_result(StoreQueryResult& res, int index)
 	// Figure out the widths of the result set's columns
 	IntVectorType widths;
 	size_t size = res.num_fields();
-	for (int i = 0; i < size; i++) {
+	for (size_t i = 0; i < size; i++) {
 		widths.push_back(max(
 				res.field(i).max_length(),
 				res.field_name(i).size()));
